PRG-53 Prefer stack over brs, add a11y label, refactor export
