+++ frontend/src/metabase/account/profile/components/UserProfileForm/UserProfileForm.tsx
@@ -9,7 +9,7 @@ import FormSelect from "metabase/core/components/FormSelect";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
 import { Form, FormProvider } from "metabase/forms";
 import * as Errors from "metabase/lib/errors";
-import { CommunityLocalizationNotice } from "metabase/localization/CommunityLocalizationNotice";
+import { CommunityLocalizationNotice } from "metabase/localization";
 import type { LocaleData, User } from "metabase-types/api";
 
 import type { UserProfileData } from "../../types";
