+++ frontend/src/metabase/admin/settings/components/SettingsPages/LocalizationSettingsPage.tsx
@@ -3,7 +3,7 @@ import _ from "underscore";
 
 import { useSetting } from "metabase/common/hooks";
 import { useSelector } from "metabase/lib/redux";
-import { CommunityLocalizationNotice } from "metabase/localization/CommunityLocalizationNotice";
+import { CommunityLocalizationNotice } from "metabase/localization";
 import { getApplicationName } from "metabase/selectors/whitelabel";
 import { Box, Stack } from "metabase/ui";
 
@@ -26,12 +26,10 @@ export function LocalizationSettingsPage() {
           )}
           inputType="select"
           description={
-            <>
+            <Stack gap="md">
               {t`The default language for all users across the ${applicationName} UI, system emails, subscriptions, and alerts. Each user can override this from their own account settings.`}
-              <br />
-              <br />
               <CommunityLocalizationNotice isAdminView />
-            </>
+            </Stack>
           }
         />
         <AdminSettingInput
