+++ frontend/src/metabase/localization/index.ts
@@ -1,0 +1,4 @@
+export {
+  getLocalizationNoticeText,
+  CommunityLocalizationNotice,
+} from "./CommunityLocalizationNotice";
